Renaming files, reworking check for compatible classifiers/regressors
