+++ src/skmatter/decomposition/_pcovr.py
@@ -1,35 +1,29 @@
-import numbers
-import warnings
-
 import numpy as np
-from numpy.linalg import LinAlgError
-from scipy import linalg
-from scipy.linalg import sqrtm as MatrixSqrt
-from scipy.sparse.linalg import svds
-from sklearn.decomposition._base import _BasePCA
-from sklearn.decomposition._pca import _infer_dimension
-from sklearn.linear_model import LinearRegression, Ridge, RidgeCV
-from sklearn.linear_model._base import LinearModel
-from sklearn.utils import check_array, check_random_state
-from sklearn.utils._arpack import _init_arpack_v0
-from sklearn.utils.extmath import randomized_svd, stable_cumsum, svd_flip
-from sklearn.utils.validation import check_is_fitted, validate_data
 
-from skmatter.utils import check_lr_fit, pcovr_covariance, pcovr_kernel
+from sklearn.base import check_X_y, check_array
+from sklearn.linear_model import (
+    LinearRegression, 
+    Ridge, 
+    RidgeCV
+)
+from sklearn.utils.validation import check_is_fitted
 
+import sys
+sys.path.append('scikit-matter')
+from src.skmatter.decomposition._pcov import _BasePCov
+from src.skmatter.utils._pcovr_utils import check_lr_fit
 
-class PCovR(_BasePCA, LinearModel):
+class PCovR(_BasePCov):
     r"""Principal Covariates Regression, as described in [deJong1992]_
     determines a latent-space projection :math:`\mathbf{T}` which
     minimizes a combined loss in supervised and unsupervised tasks.
 
     This projection is determined by the eigendecomposition of a modified gram
     matrix :math:`\mathbf{\tilde{K}}`
-
     .. math::
       \mathbf{\tilde{K}} = \alpha \mathbf{X} \mathbf{X}^T +
             (1 - \alpha) \mathbf{\hat{Y}}\mathbf{\hat{Y}}^T
-
+    
     where :math:`\alpha` is a mixing parameter and
     :math:`\mathbf{X}` and :math:`\mathbf{\hat{Y}}` are matrices of shapes
     :math:`(n_{samples}, n_{features})` and :math:`(n_{samples}, n_{properties})`,
@@ -67,11 +61,12 @@ class PCovR(_BasePCA, LinearModel):
     mixing: float, default=0.5
         mixing parameter, as described in PCovR as :math:`{\alpha}`, here named to avoid
         confusion with regularization parameter `alpha`
+
     n_components : int, float or str, default=None
         Number of components to keep.
         if n_components is not set all components are kept::
-
             n_components == min(n_samples, n_features)
+
     svd_solver : {'auto', 'full', 'arpack', 'randomized'}, default='auto'
         If auto :
             The solver is selected by a default policy based on `X.shape` and
@@ -88,13 +83,16 @@ class PCovR(_BasePCA, LinearModel):
             min(X.shape)
         If randomized :
             run randomized SVD by the method of Halko et al.
+
     tol : float, default=1e-12
         Tolerance for singular values computed by svd_solver == 'arpack'. Must be of
         range [0.0, infinity).
+
     space: {'feature', 'sample', 'auto'}, default='auto'
         whether to compute the PCovR in `sample` or `feature` space default=`sample`
         when :math:`{n_{samples} < n_{features}}` and `feature` when
         :math:`{n_{features} < n_{samples}}`
+
     regressor: {`Ridge`, `RidgeCV`, `LinearRegression`, `precomputed`}, default=None
         regressor for computing approximated :math:`{\mathbf{\hat{Y}}}`. The regressor
         should be one `sklearn.linear_model.Ridge`, `sklearn.linear_model.RidgeCV`, or
@@ -108,45 +106,55 @@ class PCovR(_BasePCA, LinearModel):
         regressed form of the targets :math:`{\mathbf{\hat{Y}}}`. If None,
         ``sklearn.linear_model.Ridge('alpha':1e-6, 'fit_intercept':False, 'tol':1e-12)``
         is used as the regressor.
+
     iterated_power : int or 'auto', default='auto'
          Number of iterations for the power method computed by svd_solver ==
          'randomized'. Must be of range [0, infinity).
+
     random_state : int, :class:`numpy.random.RandomState` instance or None, default=None
          Used when the 'arpack' or 'randomized' solvers are used. Pass an int for
          reproducible results across multiple function calls.
-    whiten : bool, deprecated
+    
+    whiten : boolean, deprecated
 
     Attributes
     ----------
     mixing: float, default=0.5
         mixing parameter, as described in PCovR as :math:`{\alpha}`
+
     tol: float, default=1e-12
         Tolerance for singular values computed by svd_solver == 'arpack'.
         Must be of range [0.0, infinity).
+
     space: {'feature', 'sample', 'auto'}, default='auto'
         whether to compute the PCovR in `sample` or `feature` space default=`sample`
         when :math:`{n_{samples} < n_{features}}` and `feature` when
         :math:`{n_{features} < n_{samples}}`
+
     n_components_ : int
         The estimated number of components, which equals the parameter n_components, or
         the lesser value of n_features and n_samples if n_components is None.
+
     pxt_ : numpy.ndarray of size :math:`({n_{samples}, n_{components}})`
         the projector, or weights, from the input space :math:`\mathbf{X}` to the
         latent-space projection :math:`\mathbf{T}`
+
     pty_ : numpy.ndarray of size :math:`({n_{components}, n_{properties}})`
         the projector, or weights, from the latent-space projection :math:`\mathbf{T}`
         to the properties :math:`\mathbf{Y}`
+
     pxy_ : numpy.ndarray of size :math:`({n_{samples}, n_{properties}})`
         the projector, or weights, from the input space :math:`\mathbf{X}` to the
         properties :math:`\mathbf{Y}`
+
     explained_variance_ : numpy.ndarray of shape (n_components,)
         The amount of variance explained by each of the selected components.
-
         Equal to n_components largest eigenvalues
         of the PCovR-modified covariance matrix of :math:`\mathbf{X}`.
+        
     singular_values_ : numpy.ndarray of shape (n_components,)
         The singular values corresponding to each of the selected components.
-
+    
     Examples
     --------
     >>> import numpy as np
@@ -167,7 +175,6 @@ class PCovR(_BasePCA, LinearModel):
            [ 0.98166504, -4.98307078],
            [-2.9963189 ,  1.98238856]])
     """
-
     def __init__(
         self,
         mixing=0.5,
@@ -178,18 +185,18 @@ class PCovR(_BasePCA, LinearModel):
         regressor=None,
         iterated_power="auto",
         random_state=None,
-        whiten=False,
+        whiten=False
     ):
-        self.mixing = mixing
-        self.n_components = n_components
-        self.space = space
-
-        self.whiten = whiten
-        self.svd_solver = svd_solver
-        self.tol = tol
-        self.iterated_power = iterated_power
-        self.random_state = random_state
-
+        super().__init__(
+            mixing=mixing,
+            n_components=n_components,
+            svd_solver=svd_solver,
+            tol=tol,
+            space=space,
+            iterated_power=iterated_power,
+            random_state=random_state,
+            whiten=whiten
+        )
         self.regressor = regressor
 
     def fit(self, X, Y, W=None):
@@ -206,6 +213,7 @@ class PCovR(_BasePCA, LinearModel):
             means and scaled. If features are related, the matrix should be scaled
             to have unit variance, otherwise :math:`\mathbf{X}` should be
             scaled so that each feature has a variance of 1 / n_features.
+
         Y : numpy.ndarray, shape (n_samples, n_properties)
             Training data, where n_samples is the number of samples and n_properties is
             the number of properties
@@ -217,56 +225,27 @@ class PCovR(_BasePCA, LinearModel):
 
             If the passed regressor = `precomputed`, it is assumed that Y is the
             regressed form of the properties, :math:`{\mathbf{\hat{Y}}}`.
+
         W : numpy.ndarray, shape (n_features, n_properties)
             Regression weights, optional when regressor=`precomputed`. If not
             passed, it is assumed that `W = np.linalg.lstsq(X, Y, self.tol)[0]`
         """
         X, Y = validate_data(self, X, Y, y_numeric=True, multi_output=True)
+        super()._fit_utils(X, Y)
 
-        # saved for inverse transformations from the latent space,
-        # should be zero in the case that the features have been properly centered
-        self.mean_ = np.mean(X, axis=0)
+        compatible_regressors = (
+            LinearRegression,
+            Ridge,
+            RidgeCV
+        )
 
-        if np.max(np.abs(self.mean_)) > self.tol:
-            warnings.warn(
-                "This class does not automatically center data, and your data mean is"
-                " greater than the supplied tolerance.",
-                stacklevel=1,
-            )
-
-        if self.space is not None and self.space not in [
-            "feature",
-            "sample",
-            "auto",
-        ]:
-            raise ValueError("Only feature and sample space are supported.")
-
-        # Handle self.n_components==None
-        if self.n_components is None:
-            if self.svd_solver != "arpack":
-                self.n_components_ = min(X.shape)
-            else:
-                self.n_components_ = min(X.shape) - 1
-        else:
-            self.n_components_ = self.n_components
-
-        if not any(
-            [
-                self.regressor is None,
-                self.regressor == "precomputed",
-                isinstance(
-                    self.regressor, 
-                    (
-                           LinearRegression,
-                           Ridge,
-                           RidgeCV
-                        ),
-                ),
-            ]
+        if self.regressor not in ["precomputed", None] and not isinstance(
+            self.regressor, compatible_regressors
         ):
             raise ValueError(
-                "Regressor must be an instance of "
-                "`LinearRegression`, `Ridge`, `RidgeCV`, or `precomputed`"
+                "Regressor must be an instance of `"
+                f"{'`, `'.join(r.__name__ for r in compatible_regressors)}`"
+                ", or `precomputed`"
             )
         
         # Assign the default regressor
@@ -280,7 +259,7 @@ class PCovR(_BasePCA, LinearModel):
             else:
                 regressor = self.regressor
 
-            self.regressor_ = check_lr_fit(regressor, X, y=Y)
+            self.regressor_ = check_lr_fit(regressor, X, Y)
 
             W = self.regressor_.coef_.T.reshape(X.shape[1], -1)
             Yhat = self.regressor_.predict(X).reshape(X.shape[0], -1)
@@ -288,26 +267,6 @@ class PCovR(_BasePCA, LinearModel):
             Yhat = Y.copy()
             if W is None:
                 W = np.linalg.lstsq(X, Yhat, self.tol)[0]
-
-        # Handle svd_solver
-        self.fit_svd_solver_ = self.svd_solver
-        if self.fit_svd_solver_ == "auto":
-            # Small problem or self.n_components_ == 'mle', just call full PCA
-            if max(X.shape) <= 500 or self.n_components_ == "mle":
-                self.fit_svd_solver_ = "full"
-            elif self.n_components_ >= 1 and self.n_components_ < 0.8 * min(X.shape):
-                self.fit_svd_solver_ = "randomized"
-            # This is also the case of self.n_components_ in (0,1)
-            else:
-                self.fit_svd_solver_ = "full"
-
-        self.n_samples_in_, self.n_features_in_ = X.shape
-        self.space_ = self.space
-        if self.space_ is None or self.space_ == "auto":
-            if self.n_samples_in_ > self.n_features_in_:
-                self.space_ = "feature"
-            else:
-                self.space_ = "sample"
 
         if self.space_ == "feature":
             self._fit_feature_space(X, Y.reshape(Yhat.shape), Yhat)
@@ -324,8 +283,6 @@ class PCovR(_BasePCA, LinearModel):
             )
 
         self.components_ = self.pxt_.T  # for sklearn compatibility
-        print("PCovR Self.pxt_ "+ str((self.pxt_).shape))
-
         return self
 
     def _fit_feature_space(self, X, Y, Yhat):
@@ -356,41 +313,7 @@ class PCovR(_BasePCA, LinearModel):
                                \mathbf{X})^{-\frac{1}{2}} \mathbf{X}^T
                                \mathbf{Y}
         """
-        Ct, iCsqrt = pcovr_covariance(
-            mixing=self.mixing,
-            X=X,
-            Y=Yhat,
-            rcond=self.tol,
-            return_isqrt=True,
-        )
-        try:
-            Csqrt = np.linalg.lstsq(iCsqrt, np.eye(len(iCsqrt)), rcond=None)[0]
-
-        # if we can avoid recomputing Csqrt, we should, but sometimes we
-        # run into a singular matrix, which is what we do here
-        except LinAlgError:
-            Csqrt = np.real(MatrixSqrt(X.T @ X))
-
-        if self.fit_svd_solver_ == "full":
-            U, S, Vt = self._decompose_full(Ct)
-        elif self.fit_svd_solver_ in ["arpack", "randomized"]:
-            U, S, Vt = self._decompose_truncated(Ct)
-        else:
-            raise ValueError(
-                "Unrecognized svd_solver='{0}'" "".format(self.fit_svd_solver_)
-            )
-
-        self.singular_values_ = np.sqrt(S.copy())
-        self.explained_variance_ = S / (X.shape[0] - 1)
-        self.explained_variance_ratio_ = (
-            self.explained_variance_ / self.explained_variance_.sum()
-        )
-
-        S_sqrt = np.diagflat([np.sqrt(s) if s > self.tol else 0.0 for s in S])
-        S_sqrt_inv = np.diagflat([1.0 / np.sqrt(s) if s > self.tol else 0.0 for s in S])
-        self.pxt_ = np.linalg.multi_dot([iCsqrt, Vt.T, S_sqrt])
-        self.ptx_ = np.linalg.multi_dot([S_sqrt_inv, Vt, Csqrt])
-        self.pty_ = np.linalg.multi_dot([S_sqrt_inv, Vt, iCsqrt, X.T, Y])
+        return super()._fit_feature_space(X, Y, Yhat)
 
     def _fit_sample_space(self, X, Y, Yhat, W):
         r"""In sample-space PCovR, the projectors are determined by:
@@ -415,144 +338,7 @@ class PCovR(_BasePCA, LinearModel):
             \mathbf{P}_{TY} = \mathbf{\Lambda}_\mathbf{\tilde{K}}^{-\frac{1}{2}}
                                \mathbf{U}_\mathbf{\tilde{K}}^T \mathbf{Y}
         """
-        Kt = pcovr_kernel(mixing=self.mixing, X=X, Y=Yhat)
-
-        if self.fit_svd_solver_ == "full":
-            U, S, Vt = self._decompose_full(Kt)
-        elif self.fit_svd_solver_ in ["arpack", "randomized"]:
-            U, S, Vt = self._decompose_truncated(Kt)
-        else:
-            raise ValueError(
-                "Unrecognized svd_solver='{0}'" "".format(self.fit_svd_solver_)
-            )
-
-        self.singular_values_ = np.sqrt(S.copy())
-        self.explained_variance_ = S / (X.shape[0] - 1)
-        self.explained_variance_ratio_ = (
-            self.explained_variance_ / self.explained_variance_.sum()
-        )
-
-        P = (self.mixing * X.T) + (1.0 - self.mixing) * W @ Yhat.T
-        S_sqrt_inv = np.diagflat([1.0 / np.sqrt(s) if s > self.tol else 0.0 for s in S])
-        T = Vt.T @ S_sqrt_inv
-
-        self.pxt_ = P @ T
-        self.pty_ = T.T @ Y
-        self.ptx_ = T.T @ X
-
-    def _decompose_truncated(self, mat):
-        if not 1 <= self.n_components_ <= min(self.n_samples_in_, self.n_features_in_):
-            raise ValueError(
-                "n_components=%r must be between 1 and "
-                "min(n_samples, n_features)=%r with "
-                "svd_solver='%s'"
-                % (
-                    self.n_components_,
-                    min(self.n_samples_in_, self.n_features_in_),
-                    self.svd_solver,
-                )
-            )
-        elif not isinstance(self.n_components_, numbers.Integral):
-            raise ValueError(
-                "n_components=%r must be of type int "
-                "when greater than or equal to 1, was of type=%r"
-                % (self.n_components_, type(self.n_components_))
-            )
-        elif self.svd_solver == "arpack" and self.n_components_ == min(
-            self.n_samples_in_, self.n_features_in_
-        ):
-            raise ValueError(
-                "n_components=%r must be strictly less than "
-                "min(n_samples, n_features)=%r with "
-                "svd_solver='%s'"
-                % (
-                    self.n_components_,
-                    min(self.n_samples_in_, self.n_features_in_),
-                    self.svd_solver,
-                )
-            )
-
-        random_state = check_random_state(self.random_state)
-
-        if self.fit_svd_solver_ == "arpack":
-            v0 = _init_arpack_v0(min(mat.shape), random_state)
-            U, S, Vt = svds(mat, k=self.n_components_, tol=self.tol, v0=v0)
-            # svds doesn't abide by scipy.linalg.svd/randomized_svd
-            # conventions, so reverse its outputs.
-            S = S[::-1]
-            # flip eigenvectors' sign to enforce deterministic output
-            U, Vt = svd_flip(U[:, ::-1], Vt[::-1])
-
-        # We have already eliminated all other solvers, so this must be "randomized"
-        else:
-            # sign flipping is done inside
-            U, S, Vt = randomized_svd(
-                mat,
-                n_components=self.n_components_,
-                n_iter=self.iterated_power,
-                flip_sign=True,
-                random_state=random_state,
-            )
-
-        return U, S, Vt
-
-    def _decompose_full(self, mat):
-        if self.n_components_ == "mle":
-            if self.n_samples_in_ < self.n_features_in_:
-                raise ValueError(
-                    "n_components='mle' is only supported " "if n_samples >= n_features"
-                )
-        elif (
-            not 0 <= self.n_components_ <= min(self.n_samples_in_, self.n_features_in_)
-        ):
-            raise ValueError(
-                "n_components=%r must be between 1 and "
-                "min(n_samples, n_features)=%r with "
-                "svd_solver='%s'"
-                % (
-                    self.n_components_,
-                    min(self.n_samples_in_, self.n_features_in_),
-                    self.svd_solver,
-                )
-            )
-        elif self.n_components_ >= 1:
-            if not isinstance(self.n_components_, numbers.Integral):
-                raise ValueError(
-                    "n_components=%r must be of type int "
-                    "when greater than or equal to 1, "
-                    "was of type=%r" % (self.n_components_, type(self.n_components_))
-                )
-
-        U, S, Vt = linalg.svd(mat, full_matrices=False)
-
-        # flip eigenvectors' sign to enforce deterministic output
-        U, Vt = svd_flip(U, Vt)
-
-        # Get variance explained by singular values
-        explained_variance_ = S / (self.n_samples_in_ - 1)
-        total_var = explained_variance_.sum()
-        explained_variance_ratio_ = explained_variance_ / total_var
-
-        # Postprocess the number of components required
-        if self.n_components_ == "mle":
-            self.n_components_ = _infer_dimension(
-                explained_variance_, self.n_samples_in_
-            )
-        elif 0 < self.n_components_ < 1.0:
-            # number of components for which the cumulated explained
-            # variance percentage is superior to the desired threshold
-            # side='right' ensures that number of features selected
-            # their variance is always greater than self.n_components_ float
-            # passed. More discussion in issue: #15669
-            ratio_cumsum = stable_cumsum(explained_variance_ratio_)
-            self.n_components_ = (
-                np.searchsorted(ratio_cumsum, self.n_components_, side="right") + 1
-            )
-        return (
-            U[:, : self.n_components_],
-            S[: self.n_components_],
-            Vt[: self.n_components_],
-        )
+        return super()._fit_sample_space(X, Y, Yhat, W)
 
     def inverse_transform(self, T):
         r"""Transform data back to its original space.
@@ -571,15 +357,7 @@ class PCovR(_BasePCA, LinearModel):
         -------
         X_original ndarray, shape (n_samples, n_features)
         """
-        if np.max(np.abs(self.mean_)) > self.tol:
-            warnings.warn(
-                "This class does not automatically un-center data, and your data mean "
-                "is greater than the supplied tolerance, so the inverse transformation "
-                "will be off by the original data mean.",
-                stacklevel=1,
-            )
-
-        return T @ self.ptx_
+        return super().inverse_transform(T)
 
     def predict(self, X=None, T=None):
         """Predicts the property values using regression on X or T."""
@@ -607,8 +385,6 @@ class PCovR(_BasePCA, LinearModel):
             New data, where n_samples is the number of samples
             and n_features is the number of features.
         """
-        check_is_fitted(self, ["pxt_", "mean_"])
-
         return super().transform(X)
 
     def score(self, X, y, T=None):
