+++ src/skmatter/decomposition/playground.py
@@ -6,10 +6,12 @@ from sklearn.discriminant_analysis import StandardScaler
 from sklearn.exceptions import NotFittedError
 from sklearn.kernel_ridge import KernelRidge
 from sklearn.linear_model import LogisticRegression, LinearRegression
+from sklearn.multioutput import MultiOutputClassifier
+from sklearn.naive_bayes import GaussianNB
 from sklearn.svm import SVC
-from kernel_pcovc_new import KernelPCovC
+from _kernel_pcovc import KernelPCovC
 from _kernel_pcovr import KernelPCovR
-from pcovc_new import PCovC
+from _pcovc import PCovC
 from sklearn.datasets import load_breast_cancer as get_dataset
 from sklearn.datasets import load_iris as get_dataset2
 from sklearn.datasets import load_diabetes as get_dataset3
@@ -26,22 +28,32 @@ X_scaled = scaler.fit_transform(X)
 # y_pred = ke.predict(X)
 # print(ke.decision_function(X))
 
-model = KernelPCovC(mixing=0.5, center=False, kernel="linear", classifier=LogisticRegression(), n_components=2)
+Y = np.column_stack((Y, Y, Y))
+print(Y.shape)
+
+model = MultiOutputClassifier(LogisticRegression())
 model.fit(X_scaled, Y)
-print(model.n_features_in_)
-T = model.transform(X_scaled)
+print(model.predict(X_scaled))
 
-Z = model.decision_function(X_scaled)
-X = model.inverse_transform(T)
-print(T.shape)
-y_pred = model.predict(X_scaled)
-print(model.score(X_scaled, Y)) # we should have KPCovC match PCovC decision function shape 
-
-model2 = PCovC(mixing=0.5, classifier=LogisticRegression(), n_components=2)
+model2 = PCovC(mixing=0.5, classifier=MultiOutputClassifier(LogisticRegression()), n_components=2)
 model2.fit(X_scaled, Y)
-T_2 = model2.transform(X_scaled)
-y_pred_2 = model2.predict(X_scaled)
-print(model2.score(X_scaled, Y))
+
+# model = KernelPCovC(mixing=0.5, center=False, kernel="linear", classifier=LogisticRegression(), n_components=2)
+# model.fit(X_scaled, Y)
+# print(model.n_features_in_)
+# T = model.transform(X_scaled)
+
+# Z = model.decision_function(X_scaled)
+# X = model.inverse_transform(T)
+# print(T.shape)
+# y_pred = model.predict(X_scaled)
+# print(model.score(X_scaled, Y)) # we should have KPCovC match PCovC decision function shape 
+
+# model2 = PCovC(mixing=0.5, classifier=LogisticRegression(), n_components=2)
+# model2.fit(X_scaled, Y)
+# T_2 = model2.transform(X_scaled)
+# y_pred_2 = model2.predict(X_scaled)
+# print(model2.score(X_scaled, Y))
 
 # ke = KernelPCovC(mixing=1.0, classifier=SVC(verbose=1), svd_solver="full",n_components=2)
 # ke.fit(X, Y)
