+++ tests/test_pcovr.py
@@ -12,7 +12,7 @@ from sklearn.utils.validation import check_X_y
 
 import sys
 sys.path.append('scikit-matter')
-from src.skmatter.decomposition.pcovr_new import PCovR
+from src.skmatter.decomposition._pcovr import PCovR
 
 class PCovRBaseTest(unittest.TestCase):
     def __init__(self, *args, **kwargs):
